`@remotion/browser-studio`: Support saving sequence props
