+++ packages/studio/src/components/sequence-props-api.ts
@@ -1,10 +1,21 @@
 import type {
+	SaveSequencePropsRequest,
+	SaveSequencePropsResponse,
 	SubscribeToSequencePropsRequest,
 	SubscribeToSequencePropsResponse,
 	UnsubscribeFromSequencePropsRequest,
 } from '@remotion/studio-shared';
 import {getBrowserStudioOperations} from '../helpers/browser-studio-operations';
 import {callApi} from './call-api';
+
+export const saveSequenceProps = (
+	request: SaveSequencePropsRequest,
+): Promise<SaveSequencePropsResponse> => {
+	const browserStudioOperations = getBrowserStudioOperations();
+	return browserStudioOperations
+		? browserStudioOperations.saveSequenceProps(request)
+		: callApi('/api/save-sequence-props', request);
+};
 
 export const subscribeToSequenceProps = (
 	request: SubscribeToSequencePropsRequest,
